refactor(platform): move base classes into platform/base sub-package
