+++ .buildkite/cases/nightly-gate-cases.txt
@@ -1,0 +1,13 @@
+# Nightly vLLM x LMCache integration gate case registry.
+#
+# Format (one case per line, '#' starts a comment, blank lines ignored):
+#   <case_name> <gpu_count> <buildkite_step_label>
+#
+# These cases run BEFORE the nightly build/wheels jobs and act as a gate:
+# if any line below fails, today's vllm-nightly is considered broken and
+# nightly artifacts are NOT published.
+#
+# To add a CPU-only gate (future PR): append a row with gpu_count=0 and
+# add a matching step in nightly-gate-pipeline.yml.
+
+cache_stats        1   gate-mp-cache-stats
