+++ lmcache/v1/platform/cuda/ipc_wrapper.py
@@ -24,7 +24,7 @@ import torch
 
 # First Party
 from lmcache import torch_device_type
-from lmcache.v1.platform.base_ipc_wrapper import DeviceIPCWrapper
+from lmcache.v1.platform.base.ipc_wrapper import DeviceIPCWrapper
 
 
 class CudaIPCWrapper(DeviceIPCWrapper):
