+++ lmcache/v1/platform/cuda/pin_memory.py
@@ -7,7 +7,7 @@ import ctypes.util
 
 # First Party
 from lmcache.logging import init_logger
-from lmcache.v1.platform.base_pin_memory import PinMemoryBackend
+from lmcache.v1.platform.base.pin_memory import PinMemoryBackend
 
 logger = init_logger(__name__)
